requirements.txt
Discord.py
openai
python-dotenv